requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Jinja2==3.0.3
click==8.0.4
itsdangerous==2.0.1
google-api-python-client==2.86.0
google-auth==2.38.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-storage==2.14.0
pytz
